Erstattet testene i vurderingUtpekt.test.js med en snapshot test, flyttet komponenten inn i egen mappe
