+++ src/sider/eu_eøs/vurderutpeking/stegMap/vurderutpeking.js
@@ -2,7 +2,7 @@ import MKV from "../../../../melosyskodeverk";
 
 import Steg from "../../../../felleskomponenter/stegvelger/stegMotor/steg";
 import { FANE_STATUS, STEG } from "../../../../felleskomponenter/stegvelger";
-import VurderingUtpekt from "../../stegKomponenter/vurderingUtpekt";
+import VurderingUtpekt from "../../stegKomponenter/vurderingUtpekt/vurderingUtpekt";
 
 import { hentLovvalgsbestemmelse } from "../../../../domeneUtils";
